# Conflicts:
#	app/src/main/java/com/namoadigital/prj001/sql/GeOsDeviceItem_Sql_003.kt
    - Criada GeOsDeviceItem_Sql_004 para UPDATE de comentario.
#	app/src/main/java/com/namoadigital/prj001/ui/act011/Act011_Main_Presenter.java
#	app/src/main/java/com/namoadigital/prj001/ui/act011/Act011_Main_Presenter_Impl.java

+++ app/src/main/java/com/namoadigital/prj001/ui/act086/Act086ProductEditDialog.kt
@@ -80,7 +80,7 @@ class Act086ProductEditDialog : BottomSheetDialogFragment() {
                 dismiss()
             }
             act086ProductEditDialogBtnApply.setOnClickListener {
-                materialItem.productQty = act086ProductEditDialogEtQty.text.toString().trim().toInt()
+                materialItem.productQty = act086ProductEditDialogEtQty.text.toString().trim().toFloat()
                 //esconde teclado ao selecionar item
                 ToolBox_Inf.hideSoftKeyboard(context,act086ProductEditDialogEtQty)
                 onApplyClick(productIdx,materialItem,isAddAction)
